image_classifier/requirements.txt
absl-py==0.6.1
bleach==1.5.0
coremltools==2.0
h5py==2.8.0
html5lib==0.9999999
Keras==2.1.6
Markdown==3.0.1
numpy==1.15.4
Pillow==5.3.0
protobuf==3.6.1
PyYAML==3.13
scipy==1.1.0
six==1.11.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
Werkzeug==0.14.1
